requirements.txt
Flask>=2.0.0
scikit-learn>=1.0
Werkzeug>=2.0
PyPDF2>=3.0.0
python-docx>=0.8.11